Refactor BinstarSpec
