+++ conda_env/specs/__init__.py
@@ -1,0 +1,4 @@
+from binstar import BinstarSpec
+
+
+all_specs = [BinstarSpec]
